+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Added a basic level up dialog
 - Removed ability score change flags (0 Str, 1 Cha, etc.)
+- Always submit data on closing an actor or item sheet
 
 ### API